+++ libp2p/src/builder/phase/tcp.rs
@@ -97,7 +97,6 @@ macro_rules! impl_tcp_builder {
     };
 }
 
-impl_tcp_builder!("async-std", super::provider::AsyncStd, async_io);
 impl_tcp_builder!("tokio", super::provider::Tokio, tokio);
 
 impl<Provider> SwarmBuilder<Provider, TcpPhase> {
@@ -218,9 +217,7 @@ macro_rules! impl_tcp_phase_with_websocket {
 impl_tcp_phase_with_websocket!(
     "async-std",
     super::provider::AsyncStd,
-    rw_stream_sink::RwStreamSink<
-        libp2p_websocket::BytesConnection<libp2p_tcp::async_io::TcpStream>,
-    >
+    rw_stream_sink::RwStreamSink<libp2p_websocket::BytesConnection<libp2p_tcp::tokio::TcpStream>>
 );
 impl_tcp_phase_with_websocket!(
     "tokio",